+++ CHANGELOG.md
@@ -2,6 +2,22 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [0.29.0-rc.94](https://github.com/SAP/fundamental-ngx/compare/v0.29.0-rc.93...v0.29.0-rc.94) (2021-04-15)
+
+## [0.29.0-rc.93](https://github.com/SAP/fundamental-ngx/compare/v0.29.0-rc.91...v0.29.0-rc.93) (2021-04-15)
+
+
+### Bug Fixes
+
+* (core) fix dialog default cancel button ([c1b4eaf](https://github.com/SAP/fundamental-ngx/commit/c1b4eafbe123aea30a5adedc3ac86a4b9b44648a))
+
+## [0.29.0-rc.92](https://github.com/SAP/fundamental-ngx/compare/v0.29.0-rc.91...v0.29.0-rc.92) (2021-04-15)
+
+
+### Bug Fixes
+
+* (core) fix dialog default cancel button ([c1b4eaf](https://github.com/SAP/fundamental-ngx/commit/c1b4eafbe123aea30a5adedc3ac86a4b9b44648a))
+
 ## [0.29.0-rc.91](https://github.com/SAP/fundamental-ngx/compare/v0.29.0-rc.90...v0.29.0-rc.91) (2021-04-14)